[-] Remove manual texture clamping and scaling since DirectX9 handles all of that itself now. [-] Remove dumping of bitmap textures [~] Slightly change the way we are handling going to fullscreen, still  needs alot of work [~] Change some uneeded LONG's to int's TODO Make it use DirectX's inbuilt function to load in the textures

+++ Config.cpp
@@ -805,7 +805,7 @@ void GenerateCurrentRomOptions()
 
 void Ini_GetRomOptions(LPGAMESETTING pGameSetting)
 {
-	LONG i;
+	int i;
 
 	i = FindIniEntry(pGameSetting->romheader.dwCRC1,
 							  pGameSetting->romheader.dwCRC2,
@@ -843,7 +843,7 @@ void Ini_GetRomOptions(LPGAMESETTING pGameSetting)
 
 void Ini_StoreRomOptions(LPGAMESETTING pGameSetting)
 {
-	LONG i;
+	int i;
 
 	i = FindIniEntry(pGameSetting->romheader.dwCRC1,
 		pGameSetting->romheader.dwCRC2,
